+++ src/containerapp/HISTORY.rst
@@ -7,6 +7,7 @@ Upcoming
 ++++++
 * 'az containerapp job start': update start execution payload format to exlude template property from API version 2023-05-01 onwards
 * 'az containerapp service': add support for creation and deletion of MariaDB
+* 'az containerapp create/list': support --environment-type parameter
 
 0.3.36
 ++++++